+++ vite.config.ts
@@ -59,5 +59,8 @@ export default defineConfig({
         src: "/src",
       },
     },
+    optimizeDeps: {
+      include: ['@mui/material/Tooltip', '@emotion/styled', '@mui/material/Unstable_Grid2'],
+    },
 })